Final edits!
I updated everything including pricing, session card photos, descriptions, call buttons, links, and the JotForm form (not on here). I don't know if I want to have my contact form present yet or not. A project for a future day. All there is left is to add the photos for each gallery.